Move components for side effects into side-effect-components/
