+++ src/app/index.jsx
@@ -6,9 +6,9 @@ import { UrlSync } from '@r/platform/components';
 import { AppMainPage } from './pages/AppMain';
 import AppOverlayMenu from './components/AppOverlayMenu';
 import { AppTopNav } from './components/AppTopNav';
-import { CookieSync } from './CookieSync';
-import { DomModifier } from './DomModifier';
-import { LocalStorageSync } from './LocalStorageSync';
+import { CookieSync } from './side-effect-components/CookieSync';
+import { DomModifier } from './side-effect-components/DomModifier';
+import { LocalStorageSync } from './side-effect-components/LocalStorageSync';
 
 export default class App extends React.Component {
   render() {
